fix: rework library loading behavior
The library loading behavior is changed so that the default filename is
used for each platform if tinyDTLS cannot be found under a list of
defined file paths.

This change allows tinyDTLS to be loaded in a Flutter app if it is
provided under the default file name, as the file system approach does
not work here. To keep the handling of missing tinyDTLS files
consistent, an ArgumentError that might be thrown is catched and a
custom Exception is thrown instead. This is not the recommended behavior
but it allows library users to offer DTLS support conditionally,
depending on whether tinyDTLS is available or not.